Deactivate teleport platform and illuminate the area
Deactivate Object:
After the player has stepped on TeleportA, add some code to make sure the player cannot be teleported again from that particular teleport platform. Added code to disable the collider for 2 seconds only

Illuminate Area:
When the player is teleported, turn on a light which illuminates where the player should go to next (eg. the next Teleport platform).