Teleport player to randon platform and blink world light
Blink World Light:
Use a coroutine to quickly blink the main world light on then off so that the player gets a glimpse at the whole level. The idea is that your world is quite dark but for a brief moment it is lit up clearly.

Teleport Player Random:
Instead of teleporting to just one location, randomly teleport the player to one of many locations.